update tests to check for schema instead of dataset
